Conflicts:
	Planetarium/src/it/unibs/eclisse/planetarium/PlanetarySystem.java

+++ Planetarium/src/it/unibs/eclisse/planetarium/Moon.java
@@ -2,8 +2,10 @@ package it.unibs.eclisse.planetarium;
 
 public class Moon extends CelestialBody{
 	
-	public Moon(String nome, double massa, int id, int x, int y) {
+	
+	public Moon(String nome, double massa, int id, double x, double y) {
 		
 		super(nome, massa, id, x, y);
 	}
+	
 }
